Create file for get, oneByVersion and latest
